pyproject.toml
[tool.poetry]
name = "qaguru-hw-6-21"
version = "0.1.0"
description = ""
authors = ["Торсунова Екатерина Павловна <katrin.tep@list.ru>"]
readme = "README.md"
packages = [{include = "qaguru_hw_6"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.0"
requests = "^2.30.0"
jsonschema = "4.19.0"
allure-pytest = "2.13.2"
curlify = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
